+++ nativehook/src/main/cpp/asm/instruction.h
@@ -5,8 +5,7 @@
 #ifndef SANDHOOK_INSTRUCTION_H
 #define SANDHOOK_INSTRUCTION_H
 
-#include <malloc.h>
-#include "../includes/base.h"
+#include "unit.h"
 
 //aarch64
 typedef U32 InstA64;
@@ -35,60 +34,6 @@ if (COND) { \
 
 namespace SandHook {
     namespace Asm {
-
-        template <typename Raw>
-        class Unit {
-        public:
-
-            Unit() {
-                if (unitType() != Void) {
-                    raw = reinterpret_cast<Raw *>(malloc(size()));
-                    memset(raw, 0, size());
-                    auto_alloc = true;
-                }
-            }
-
-            Unit<Raw>(Raw *raw) : raw(raw) {}
-
-            Unit<Raw>(Raw raw) {
-                Unit();
-                *this->raw = raw;
-            }
-
-            virtual void* getPC() {
-                return auto_alloc ? nullptr : raw;
-            }
-
-            inline Raw* get() const {
-                return raw;
-            }
-
-            inline void set(Raw raw) const {
-                *this->raw = raw;
-            }
-
-            inline void copy(void* dest) {
-                memcpy(dest, getPC(), size());
-            }
-
-            virtual UnitType unitType() {
-                return UnitType::Unkown;
-            };
-
-            virtual U32 size() {
-                return sizeof(Raw);
-            }
-
-            virtual ~Unit() {
-                if (auto_alloc) {
-                    free(raw);
-                }
-            }
-
-        private:
-            Raw* raw;
-            bool auto_alloc = false;
-        };
 
         template <typename Inst>
         class Instruction : public Unit<Inst> {
@@ -128,41 +73,6 @@ namespace SandHook {
 
         protected:
             bool valid = true;
-        };
-
-        template <typename DType>
-        class Data : public Unit<DType> {
-        public:
-            Data(DType raw) : Unit<DType>(raw) {}
-            inline UnitType unitType() override {
-                return UnitType::Data;
-            };
-        };
-
-        class Data16 : public Data<U16> {
-        public:
-            Data16(U16 raw) : Data(raw) {}
-        };
-
-        class Data32 : public Data<U32> {
-        public:
-            Data32(U32 raw) : Data(raw) {}
-        };
-
-        class Data64 : public Data<U64> {
-        public:
-            Data64(U64 raw) : Data(raw) {}
-        };
-
-        class Label : public Unit<None> {
-        public:
-            Label() {}
-            inline UnitType unitType() override {
-                return UnitType::Label;
-            }
-            U32 size() override {
-                return 0;
-            }
         };
 
         class Void : public Unit<None> {
